+++ Entities/AHuman.cpp
@@ -137,35 +137,6 @@ int AHuman::Create()
     return 0;
 }
 
-/*
-//////////////////////////////////////////////////////////////////////////////////////////
-// Method:          Create
-//////////////////////////////////////////////////////////////////////////////////////////
-// Description:     Makes the AHuman object ready for use.
-
-int AHuman::Create(BITMAP *pSprite,
-                   Controller *pController,
-                   const float mass,
-                   const Vector &position,
-                   const Vector &velocity,
-                   AtomGroup *hitBody,
-                   const unsigned long lifetime,
-                   Status status,
-                   const int health)
-{
-    
-
-    return Actor::Create(pSprite,
-                         pController,
-                         mass,
-                         position,
-                         velocity,
-                         hitBody,
-                         lifetime,
-                         status,
-                         health);
-}
-*/
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // Method:          Create
@@ -175,6 +146,8 @@ int AHuman::Create(BITMAP *pSprite,
 int AHuman::Create(const AHuman &reference)
 {
     Actor::Create(reference);
+
+	m_ThrowPrepTime = reference.m_ThrowPrepTime;
 
     if (reference.m_pHead) {
         m_pHead = dynamic_cast<Attachable *>(reference.m_pHead->Clone());
@@ -258,7 +231,9 @@ int AHuman::Create(const AHuman &reference)
 
 int AHuman::ReadProperty(std::string propName, Reader &reader)
 {
-    if (propName == "Head")
+	if (propName == "ThrowPrepTime")
+		reader >> m_ThrowPrepTime;
+	else if (propName == "Head")
     {
         delete m_pHead;
         m_pHead = new Attachable;
@@ -365,6 +340,8 @@ int AHuman::Save(Writer &writer) const
 {
     Actor::Save(writer);
 
+	writer.NewProperty("ThrowPrepTime");
+	writer << m_ThrowPrepTime;
     writer.NewProperty("Head");
     writer << m_pHead;
     writer.NewProperty("Jetpack");
@@ -411,47 +388,6 @@ int AHuman::Save(Writer &writer) const
     return 0;
 }
 
-/*
-//////////////////////////////////////////////////////////////////////////////////////////
-// Virtual method:  Create
-//////////////////////////////////////////////////////////////////////////////////////////
-// Description:     Makes the AHuman object ready for use.
-
-int AHuman::Create(istream &stream, bool checkType)
-{
-    if (checkType)
-    {
-        string name;
-        stream >> name;
-        if (name != m_sClass.GetName())
-        {
-           RTEAbort("Wrong type in stream when passed to Create");
-           return -1;
-        }
-    }
-
-    Actor::Create(stream);
-
-    return 0;
-}
-
-
-//////////////////////////////////////////////////////////////////////////////////////////
-// Virtual method:  Save
-//////////////////////////////////////////////////////////////////////////////////////////
-// Description:     Saves the complete state of this AHuman to an output stream for
-//                  later recreation with Create(istream &stream);
-
-int AHuman::Save(ostream &stream) const
-{
-    stream << m_sClass.GetName() << " ";
-
-    Actor::Save(stream);
-//    stream << " ";
-
-    return 0;
-}
-*/
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // Method:          Destroy